blkjk: implement simple game
The hotpath() now contains the game logic for a simple version of Blackjack :^)
hotpath() is called once a user is known to be present (WIP) & ready to play.
hotpath() will determine if the player is valid using the Enclave network library (TODO),
then it will allow the player to input a bet quantity according to their balance given by the SEP over the Enclave network.
Once a bet is determined, a game-draw repeat loop is entered using the `local playing` variable.
The main repeat loop is comprised of sub game-draw loops such as player hitting, dealer hitting, but also handles all logic flow for the game state.
The player's balance, stored in the `game.player.balance` global, is only changed after a hand is fully played out and the bet can bet settled against the balance;
this is not a problem because the place_bet() logic already handles parsing a proper bet input from the player.